requirements.txt
python-dotenv
openai-agents
openai-agents[litellm]
